fix(folder-structure): reorganized folder structure for better access
