+++ authservice/src/main/java/auth/proximity/authservice/controller/AuthController.java
@@ -1,18 +1,22 @@
 package auth.proximity.authservice.controller;
 
 import auth.proximity.authservice.dto.*;
+import auth.proximity.authservice.dto.user.AdminUpdatePasswordRequest;
+import auth.proximity.authservice.dto.user.UserDto;
+import auth.proximity.authservice.dto.user.UserInfoResponse;
+import auth.proximity.authservice.dto.user.UserUpdateRequest;
 import auth.proximity.authservice.entity.User;
 
-import auth.proximity.authservice.security.dto.LoginRequest;
-import auth.proximity.authservice.security.dto.LoginResponse;
-import auth.proximity.authservice.security.dto.RefreshTokenResponse;
-import auth.proximity.authservice.security.dto.InfoResponse;
+import auth.proximity.authservice.dto.security.LoginRequest;
+import auth.proximity.authservice.dto.security.LoginResponse;
+import auth.proximity.authservice.dto.security.RefreshTokenResponse;
+import auth.proximity.authservice.dto.security.InfoResponse;
 import auth.proximity.authservice.security.jwt.JwtConstants;
 import auth.proximity.authservice.security.jwt.JwtUtils;
-import auth.proximity.authservice.security.service.UserDetailsImpl;
-import auth.proximity.authservice.security.service.UserDetailsServiceImpl;
-import auth.proximity.authservice.service.IUserService;
-import auth.proximity.authservice.service.ProfilePictureService;
+import auth.proximity.authservice.services.security.UserDetailsImpl;
+import auth.proximity.authservice.services.security.UserDetailsServiceImpl;
+import auth.proximity.authservice.services.IUserService;
+import auth.proximity.authservice.services.ProfilePictureService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
