+++ authservice/src/main/java/auth/proximity/authservice/dto/user/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package auth.proximity.authservice.dto;
+package auth.proximity.authservice.dto.user;
 
 import auth.proximity.authservice.entity.Role;
