+++ authservice/src/main/java/auth/proximity/authservice/dto/user/UserPasswordResetRequest.java
@@ -1,4 +1,4 @@
-package auth.proximity.authservice.dto;
+package auth.proximity.authservice.dto.user;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
