+++ authservice/src/main/java/auth/proximity/authservice/dto/user/UserUpdateRequest.java
@@ -1,5 +1,5 @@
 
-package auth.proximity.authservice.dto;
+package auth.proximity.authservice.dto.user;
 public record UserUpdateRequest(
         String userName,
         String phoneNumber,
